move to OneTimeTaskRunner:: (rename files, classes); + config initializer
